FIX / 홈 디자인 수정
