+++ lib/home/notice/noticeMain.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
+import '../../home/notice/noticeDetail.dart';
+
 class NoticeMain extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
@@ -55,7 +57,7 @@ class _NoticeMain extends State<NoticeMain> {
         backgroundColor: Colors.grey[200],
         elevation: 0,
         leading: IconButton(
-          icon: Icon(Icons.arrow_back, color: Colors.black),
+          icon: Icon(Icons.arrow_back_ios, color: Colors.black),
           iconSize: 26,
           onPressed: () => context.pop(),
         ),
@@ -90,7 +92,7 @@ class _NoticeMain extends State<NoticeMain> {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                        builder: (context) => NoticeDetailScreen(
+                        builder: (context) => NoticeDetail(
                           id: notice["id"]!,
                           title: notice["title"]!,
                           date: notice["date"]!,
@@ -211,59 +213,6 @@ class _NoticeMain extends State<NoticeMain> {
             ),
           ),
         ],
-      ),
-    );
-  }
-}
-
-
-class NoticeDetailScreen extends StatelessWidget {
-  final String id;
-  final String title;
-  final String date;
-
-  NoticeDetailScreen({required this.id, required this.title, required this.date});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("공지사항 글", style: TextStyle(color: Colors.black)),
-        backgroundColor: Colors.white,
-        elevation: 1,
-        leading: IconButton(
-          icon: Icon(Icons.arrow_back, color: Colors.black),
-          onPressed: () => Navigator.pop(context),
-        ),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              id,
-              style: TextStyle(fontSize: 14, color: Colors.grey[600]),
-            ),
-            SizedBox(height: 8),
-            Text(
-              title,
-              style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-            ),
-            SizedBox(height: 8),
-            Text(
-              date,
-              style: TextStyle(fontSize: 14, color: Colors.grey[600]),
-            ),
-            SizedBox(height: 16),
-            Expanded(
-              child: Text(
-                "공지사항 내용. 데이터베이스에서 가져올 예정",
-                style: TextStyle(fontSize: 16, color: Colors.black),
-              ),
-            ),
-          ],
-        ),
       ),
     );
   }
